+++ packages/xchain-binance/__tests__/client.test.ts
@@ -76,7 +76,7 @@ describe('BinanceClient Test', () => {
 
   const mainnetClientURL = 'https://dex.binance.org'
   const testnetClientURL = 'https://testnet-dex.binance.org'
-  const thornodeMainetClientURL = 'https://thornode.thorchain.info'
+  const thornodeMainetClientURL = 'https://thornode.ninerealms.com'
 
   beforeEach(async () => {
     bnbClient = new BinanceClient({ phrase, network: 'mainnet' as Network })