+++ packages/xchain-bitcoin/src/utils.ts
@@ -89,6 +89,7 @@ export const arrayAverage = (array: number[]): number => {
 export const btcNetwork = (network: Network): Bitcoin.Network => {
   switch (network) {
     case Network.Mainnet:
+    case Network.Stagenet:
       return Bitcoin.networks.bitcoin
     case Network.Testnet:
       return Bitcoin.networks.testnet
@@ -106,6 +107,7 @@ export const btcNetwork = (network: Network): Bitcoin.Network => {
 export const getBalance = async (params: AddressParams): Promise<Balance[]> => {
   switch (params.network) {
     case Network.Mainnet:
+    case Network.Stagenet:
       return [
         {
           asset: AssetBTC,
@@ -181,7 +183,8 @@ export const scanUTXOs = async ({
           } as UTXO),
       )
     }
-    case Network.Mainnet: {
+    case Network.Mainnet:
+    case Network.Stagenet: {
       let utxos: haskoinApi.UtxoData[] = []
 
       if (confirmedOnly) {
@@ -344,6 +347,7 @@ export const getDefaultFees = (): Fees => {
 export const getPrefix = (network: Network) => {
   switch (network) {
     case Network.Mainnet:
+    case Network.Stagenet:
       return 'bc1'
     case Network.Testnet:
       return 'tb1'