+++ packages/xchain-thorchain/__tests__/client.test.ts
@@ -153,6 +153,9 @@ describe('Client Test', () => {
 
     thorClient.setNetwork('mainnet' as Network)
     expect(thorClient.validateAddress(thorClient.getAddress())).toEqual(true)
+
+    thorClient.setNetwork('stagenet' as Network)
+    expect(thorClient.validateAddress(thorClient.getAddress())).toEqual(true)
   })
 
   it('should have right client url', async () => {
@@ -160,6 +163,10 @@ describe('Client Test', () => {
       mainnet: {
         node: 'new mainnet client',
         rpc: 'new mainnet client',
+      },
+      stagenet: {
+        node: 'new stagenet client',
+        rpc: 'new stagenet client',
       },
       testnet: {
         node: 'new testnet client',
@@ -172,6 +179,9 @@ describe('Client Test', () => {
 
     thorClient.setNetwork('testnet' as Network)
     expect(thorClient.getClientUrl().node).toEqual('new testnet client')
+
+    thorClient.setNetwork('stagenet' as Network)
+    expect(thorClient.getClientUrl().node).toEqual('new stagenet client')
   })
 
   it('returns private key', async () => {