+++ packages/xchain-thorchain/src/util.ts
@@ -98,6 +98,8 @@ export const getPrefix = (network: Network) => {
   switch (network) {
     case Network.Mainnet:
       return 'thor'
+    case Network.Stagenet:
+      return 'sthor'
     case Network.Testnet:
       return 'tthor'
   }
@@ -271,8 +273,12 @@ export const getDefaultClientUrl = (): ClientUrl => {
       node: 'https://testnet.thornode.thorchain.info',
       rpc: 'https://testnet.rpc.thorchain.info',
     },
+    [Network.Stagenet]: {
+      node: 'https://stagenet-thornode.ninerealms.com',
+      rpc: 'https://stagenet-rpc.ninerealms.com',
+    },
     [Network.Mainnet]: {
-      node: 'https://thornode.thorchain.info',
+      node: 'https://thornode.ninerealms.com',
       rpc: 'https://rpc.thorchain.info',
     },
   }
@@ -288,16 +294,19 @@ const DEFAULT_EXPLORER_URL = 'https://viewblock.io/thorchain'
 export const getDefaultExplorerUrls = (): ExplorerUrls => {
   const root: ExplorerUrl = {
     [Network.Testnet]: `${DEFAULT_EXPLORER_URL}?network=testnet`,
+    [Network.Stagenet]: `${DEFAULT_EXPLORER_URL}?network=stagenet`,
     [Network.Mainnet]: DEFAULT_EXPLORER_URL,
   }
   const txUrl = `${DEFAULT_EXPLORER_URL}/tx`
   const tx: ExplorerUrl = {
     [Network.Testnet]: txUrl,
+    [Network.Stagenet]: txUrl,
     [Network.Mainnet]: txUrl,
   }
   const addressUrl = `${DEFAULT_EXPLORER_URL}/address`
   const address: ExplorerUrl = {
     [Network.Testnet]: addressUrl,
+    [Network.Stagenet]: addressUrl,
     [Network.Mainnet]: addressUrl,
   }
 
@@ -338,6 +347,8 @@ export const getExplorerAddressUrl = ({
   switch (network) {
     case Network.Mainnet:
       return url
+    case Network.Stagenet:
+      return `${url}?network=stagenet`
     case Network.Testnet:
       return `${url}?network=testnet`
   }
@@ -364,6 +375,8 @@ export const getExplorerTxUrl = ({
   switch (network) {
     case Network.Mainnet:
       return url
+    case Network.Stagenet:
+      return `${url}?network=stagenet`
     case Network.Testnet:
       return `${url}?network=testnet`
   }